requirements.txt
flask
pyjwt
pyotp
qrcode[pil]
bcrypt
